Growing more project

+++ types.h
@@ -1,7 +1,7 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-typedef unsigned int idType;
+typedef unsigned long long idType;
 typedef unsigned int rfidType;
 typedef double balanceType;
 typedef unsigned short cardTypeT;
